@@ -10,28 +10,27 @@ environment:
   sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
-  adapty_flutter: ^1.0.14
-  animate_do: ^2.1.0
-  bloc: ^8.0.3
-  carrier_info: ^2.0.5
+  adapty_flutter: ^2.2.5
+  animate_do: ^3.0.2
+  bloc: ^8.1.1
+  carrier_info: 2.0.5
   cupertino_icons: ^1.0.5
-  cupertino_lists: ^1.0.1
   cupertino_onboarding: ^1.1.1
-  dart_ping: ^7.0.2
-  dart_ping_ios: ^2.0.1
-  dynamic_color: ^1.5.4
+  dart_ping: ^7.0.1
+  dart_ping_ios: ^2.0.0
+  dynamic_color: ^1.6.2
   enum_to_string: ^2.0.1
   equatable: ^2.0.5
-  flex_color_scheme: ^6.0.1
+  flex_color_scheme: ^6.1.2
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.0.1
+  flutter_bloc: ^8.1.2
   flutter_platform_widgets: ^2.0.0
   flutter_svg: ^1.1.5
-  font_awesome_flutter: ^10.2.1
-  google_fonts: ^3.0.1
-  google_maps_flutter: ^2.2.0
-  google_mobile_ads: ^2.0.1
+  font_awesome_flutter: ^10.3.0
+  google_fonts: ^4.0.3
+  google_maps_flutter: ^2.2.3
+  google_mobile_ads: ^2.3.0
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   http: ^0.13.5
@@ -43,16 +42,16 @@ dependencies:
   # lan_scanner: ^3.5.0
   meta: ^1.8.0
   mocktail: ^0.3.0
-  modal_bottom_sheet: ^2.1.2
-  network_info_plus: ^2.3.1
+  modal_bottom_sheet: ^3.0.0-pre
+  network_info_plus: ^3.0.1
   network_tools: ^3.0.0+2
-  package_info_plus: ^1.4.3+1
+  package_info_plus: ^3.0.2
   path: ^1.8.2
-  path_provider: ^2.0.11
-  permission_handler: ^10.1.0
-  sentry_flutter: ^6.11.2
-  url_launcher: ^6.1.6
-  wake_on_lan: ^3.0.0
+  path_provider: ^2.0.12
+  permission_handler: ^10.2.0
+  sentry_flutter: ^6.20.1
+  url_launcher: ^6.1.9
+  wake_on_lan: ^3.1.0+6
   wiredash: ^1.5.0
 
 dev_dependencies:
